v2_sheath
Made some decent I think progress with the sheath. Still need to add
more details to it then perhaps chop it up and apply the boolean
modifier on the blade to hollow it out. See if that works. Tried
applying the boolean modifier on the entire sheath and my machine choked