v2_sheath
Tons of progress across the board. The initial blade stl files printing
straight up, well, was a pain to separate from the blade. So now you can
generate the gcode from the blade file with them laying down split on
the print bed. Got the sheath almost done too. Just need to chop it up,
add wooden dowel slots, and add a place for the strap