v2 - a big change to sheath
A big improvement to the sheath cleaning it up and making it look more
symmetrical. Now if the sword fits then we should be good. I sliced up
the sheath chopping it up in four pieces. The tip is about a ten hour
print. If that seems to fit I will print the next part and I will update
the README once the full sheath has been printed